refactor: remove /v3 prefix from stargate endpoint
