+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rsk-explorer-api",
-  "version": "2.2.2",
+  "version": "2.2.3",
   "description": "RSK Explorer API",
   "main": "index.js",
   "scripts": {
