+++ src/api/routes/api.js
@@ -1,7 +1,6 @@
 import express from 'express'
 import bodyParser from 'body-parser'
-import httpLogger from '../httpLogger'
-import v3ApiRoutes from '../v3/api'
+import stargateRoutes from '../stargate/stargate.routes'
 
 const router = express.Router()
 
@@ -47,11 +46,7 @@ const Routes = ({ log, api }, send) => {
     return sendResult({ res, req, next }, req.body)
   })
 
-  // v3
-  router.use('/v3', httpLogger)
-  router.use('/v3', express.json())
-  router.use('/v3', express.urlencoded({ extended: true }))
-  router.use('/v3', v3ApiRoutes)
+  router.use('/stargate', stargateRoutes)
 
   return router
 }
