--- src/api/v3/api.js
@@ -1,14 +1,0 @@
-import express from 'express'
-import stargateRoutes from './routes/stargate.routes'
-
-const router = express.Router()
-
-router.get('/', (req, res) => {
-  res.send({
-    message: 'RSK Explorer API v3'
-  })
-})
-
-router.use('/stargate', stargateRoutes)
-
-export default router
